fix: put the version number in the shared library name
This commit creates a define with a version number. The compiled shared
library will have this number in its name. This is important because
most operating systems refuse to load a shared library that has the same
name as a shared library that is already loaded.

The commit also includes a fix for the number of loads counter that is
included in the modules private data.