feat: add alternative port-based implementation
This commit adds an alternative
[port](https://www.erlang.org/doc/tutorial/c_port.html)-based
implementation of the JQ API as an alternative to the NIF-based one. The
user can configure which implementation to use (this is described in
the `README.md` file).

The port-based implementation seems to be about 3 times slower for
simple jq filter programs (benchmark in `test/jq_tests.erl`) but it
should be safer as bugs in jq cannot cause crashes and memory leaks in
the Erlang VM. The port program can be configured to be restarted
periodically (also described in `README.md`) which could be useful if
memory leaks are later found in jq.

TODO: c_src/enif_jq.c and c_src/erlang_jq_port_process.c contains code
that has a lot of similarities. This should be fixed by refactoring out
common code to a file that is used both by the NIF library and the
port-program.