feat: add support for timeout when processing JSON with jq
This commit adds the function `jq:process_json/3` which is like
`jq:process_json/2` except that the user can also set a timeout.
If a timeout is set and the execution of the jq program has not
finished within the timeout time, the execution of the jq program
will be terminated and an error will be returned. This is an important
feature since jq programs might execute forever. This commit only
implements timeout support in the port program implementation so the
port program implementation is used whenever a timeout is set, even if
the user has configured the library to use the NIF implementation
(timeout support for the NIF implementation might be added in a later
commit).