fix: bug that could cause hanging jq port programs after Erlang exists
If the Erlang VM is brutally killed while executing long running jq
programs with the Erlang port backend of this library, the Port programs
could be left executing alone (possible forever). This commit fixes this
issue by introducing an intermediate program whose only purpose is to
forward data to the main worker program. This way there is always a
process that can detect that STDIN is closed even if the main program is
busy executing jq code.

This will make the port program less efficient as it introduces one more
indirection for data going from Erlang to the port program. However, it
fixes a serious issue and additionally makes the code for killing the
port program less ugly.

I investigated possible solutions using signal handlers to detect when a
file descriptor is closed while the program is busy executing JQ code.
However, finding such a solution that is platform independent and robust
seems to be very tricky. The main approach I tried periodically
triggered a SIGALRM signal handler and tried to check if STDIN had been
closed. Unfortunately, it seems like STDIN is not closed even though the
write end of STDIN has been closed. It seems like one have to read data
to detect if the file descriptor is closed but that is tricky to do in a
signal handler.